+++ go/plugins/vertexai/vertexai.go
@@ -307,9 +307,9 @@ func translateResponse(resp *genai.GenerateContentResponse) *ai.GenerateResponse
 	}
 	r.Usage = &ai.GenerationUsage{}
 	if u := resp.UsageMetadata; u != nil {
-		r.Usage.InputTokens = int64(u.PromptTokenCount)
-		r.Usage.OutputTokens = int64(u.CandidatesTokenCount)
-		r.Usage.TotalTokens = int64(u.TotalTokenCount)
+		r.Usage.InputTokens = int(u.PromptTokenCount)
+		r.Usage.OutputTokens = int(u.CandidatesTokenCount)
+		r.Usage.TotalTokens = int(u.TotalTokenCount)
 	}
 	return r
 }
